separate tests about election
